requirements.txt
aenum==3.1.15
aiofiles==24.1.0
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
astor==0.8.1
attrs==24.2.0
autoflake==2.3.1
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
distro==1.9.0
effdet==0.4.1
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.55.2
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.1
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
importlib_metadata==8.5.0
infi.systray==0.1.12
iopath==0.1.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langdetect==1.0.9
layoutparser==0.3.4
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
maybe-else==0.2.1
mistune==3.0.2
mpmath==1.3.0
msal==1.31.1
mypy==1.13.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
Office365-REST-Python-Client==2.5.13
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.57.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==3.0.0b1
pandas==2.2.3
pandocfilters==1.5.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.21.0
pikepdf==9.4.2
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.0.0
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pypandoc==1.14
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.12
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
types-requests==2.32.0.20241016
types-ujson==5.10.0.20240515
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.9
unstructured-api-tools==0.10.11
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured-ingest==0.3.6
unstructured.paddleocr==2.8.1.0
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
webencodings==0.5.1
websockets==14.1
wikipedia==1.4.0
wrapt==1.17.0
xlrd==2.0.1
XlsxWriter==3.2.0
zipp==3.21.0
